Move encoder to different file and pep8.
